:hammer: Refactor QR code UI for multiplatform reuse
