+++ composeApp/src/commonMain/kotlin/com/crosspaste/image/PlatformImage.kt
@@ -2,9 +2,10 @@ package com.crosspaste.image
 
 abstract class PlatformImage(
     protected val data: ByteArray,
-    protected val width: Int,
-    protected val height: Int,
 ) {
 
-    abstract fun toImage(): Any
+    abstract fun toImage(
+        width: Int,
+        height: Int,
+    ): Any
 }
